+++ caten_worship/__init__.py
@@ -5,10 +5,17 @@ import os
 
 from flask import Flask
 
-config = os.environ.get("APP_SETTING")
+def create_app(config_code):
 
+    global config
 
-def create_app():
+    if config_code == "test":
+        config = os.environ.get("TEST_SETTING")
+    elif config_code == "":
+        config = os.environ.get("APP_SETTING")
+    else:
+        print("Wrong Config Code.")
+        exit(1)
 
     app = Flask(__name__)
     app.config.from_object(config)